bus/dpaa: add ONIC port mode for the DPAA eth
The OH ports can also be used by two application, processing contexts
to communicate to each other.
This patch enables this mode for dpaa-eth OH port as ONIC port,
so that application can use the dpaa-eth to communicate to each
other on the same SoC.

Again,this properties is driven by the system device-tree variables.

Signed-off-by: Rohit Raj <rohit.raj@nxp.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>