timer: allow platform to override cpu TSC frequency
The CPU provided value is often not accurate, allow overriding it
based on info from the host OS.

On Linux X86, if the tsc_known_freq cpu flag is missing, it means
the kernel doesn't trust it and calculates its own. We should do
the same to avoid drift.

On freebsd we have access to the kernel tsc_hz value, just use it.

Signed-off-by: Isaac Boukris <iboukris@gmail.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>