test: fix option devices
Without using allow (-a) or block (-b), EAL loads all devices by default.
Unexpected devices may be loaded when running test cases in sub-processes.

This patch fixes the issue by copying the parameters of the primary process
if the allow (-a) or block (-b) option is not used when starting the child
process.

Also, EAL does not allow the options allow (-a) and block (-b) to be used
at the same time.

Fixes: b3ce7891ad38 ("test: fix probing in secondary process")
Cc: stable@dpdk.org

Signed-off-by: Mingjin Ye <mingjinx.ye@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>