net/bnxt: tf_ulp: add track type feature to tables
Added track type field to the mapper tables so that resources
can be tracked by session id or function id. By default, the track type
is session id unless specified as function id.

This patch also includes a couple of additional changes.

action template consolidation for Thor/Thor2
	Reduced the number of action templates in order to reduce
	duplication and make more permutations of actions easier to
	handle going forward.

Change on GPE flow creation/deletion
	Corrected the tunnel type for GPE flow creation;
	Corrected the port number for GPE flow deletion.

This patch also updates template files for the following list
of changes, that are being added in this patch.

- add track type feature to tables
- action template consolidation for Thor/Thor2

Signed-off-by: Kishore Padmanabha <kishore.padmanabha@broadcom.com>
Signed-off-by: Sriharsha Basavapatna <sriharsha.basavapatna@broadcom.com>
Reviewed-by: Farah Smith <farah.smith@broadcom.com>
Reviewed-by: Ajit Khaparde <ajit.khaparde@broadcom.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>