net/bnxt: tf_ulp: add feature bit support
Added feature bit support to enable support for setting default
destination mac for ingress flows that do not specify destination
mac address.

This patch also provides the below related change.

Added support for function operand size so correct size of the
data is used.

This patch also updates template files for the following list
of changes, that are being added in this patch.

Signed-off-by: Kishore Padmanabha <kishore.padmanabha@broadcom.com>
Signed-off-by: Sriharsha Basavapatna <sriharsha.basavapatna@broadcom.com>
Reviewed-by: Michael Baucom <michael.baucom@broadcom.com>
Reviewed-by: Ajit Khaparde <ajit.khaparde@broadcom.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>