igc: fix invalid length and corrupted multi-segment mbufs
The issue only appeared with hardware-timestamping enabled
(RTE_ETH_RX_OFFLOAD_TIMESTAMP).

The length of the prepended hardware timestamp was not subtracted from
the data length so that received packets were 16 bytes longer than
expected.

In scatter-gather mode only the first mbuf has a timestamp but the
data offset of the follow-up mbufs was not adjusted accordingly.
This caused 16 bytes of packet data to be missing between
the segments.

Signed-off-by: Martin Weiser <martin.weiser@allegro-packets.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>