uapi: introduce kernel uAPI headers import
This patch introduces uAPI headers import into the DPDK
repository. This import is possible thanks to Linux Kernel
licence exception for syscalls:

https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/tree/LICENSES/exceptions/Linux-syscall-note

Header files have to be explicitly imported.

Guidelines are provided in the documentation, and helper
script is also provided to ensure proper import of the
headers (unmodified content from a released Kernel
version).

Signed-off-by: Maxime Coquelin <maxime.coquelin@redhat.com>
Acked-by: Morten Brørup <mb@smartsharesystems.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>