net/ena: add histogram support
The framework computes the differences between specified
start and end points in the code, providing insights into
performance by outputting a histogram of the time intervals.
The histogram can be used to track metrics like packet
processing latency in the network driver.
This framework requires enablement via a compilation
flag and is excluded in release builds by default.

Signed-off-by: Amit Bernstein <amitbern@amazon.com>
Signed-off-by: Shai Brandes <shaibran@amazon.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>