build: output source and build dirs at end of config
The configuration run of meson produces a lot of output. When running
test-meson-builds script, or any other script which does multiple
builds, it can be useful to see the directory in which ninja is being
run. Currently this requires scrolling back to the top of the output, so
to simplify things print out the source and build directories at the end
of the config run, so they are on-screen while ninja is printing its
output immediately underneath.

Signed-off-by: Bruce Richardson <bruce.richardson@intel.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>