doc: update match with compare result item limitation
In switch mode, when ``repr_matching_en`` flag is enabled in the devarg,
the match with compare result item is not supported to the ``ingress``
rule as an implicit REPRESENTED_PORT need to be added to the matcher.
That REPRESENTED_PORT item conflicts with the single item limitation for
match with compare result item.

Fixes: cb25df7ce9d6 ("net/mlx5: support comparison matching")
Cc: stable@dpdk.org

Signed-off-by: Suanming Mou <suanmingm@nvidia.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>