doc: modernize build instructions on Windows
Meson 0.57 was an unstable version and is now outdated.
The referenced bug in Meson 0.58 is fixed in stable releases.
Recommend the first version containing the fix: 1.5.2.

Building DPDK applications that run on 32-bit Windows is
currently not supported. However, some Visual Studio environments
default to producing 32-bit binaries.
Recommend instructing the developer prompt to produce 64-bit binaries
when that is the case.

Signed-off-by: Andre Muezerie <andremue@linux.microsoft.com>
Signed-off-by: 0-day Robot <robot@bytheb.org>